move Ticker to progress_bar.rs
